+++ CHANGELOG.md
@@ -17,6 +17,7 @@ See [keep a changelog] for information about writing changes to this log.
 - Added tags filter
 - Added pagination and match filter
 - Added date filters
+- Ensure combined filters are possible
 
 [keep a changelog]: https://keepachangelog.com/en/1.1.0/
 [unreleased]: https://github.com/itk-dev/event-database-imports/compare/main...develop